+++ src/components/input/input.tsx
@@ -166,7 +166,7 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
     /**
      * If `true`, the user cannot interact with the input.
      */
-    @Prop() disabled = false;
+    @Prop({ reflect: true }) disabled = false;
 
     /**
      * A hint to the browser for which enter key to display.
@@ -242,7 +242,7 @@ export class Input implements ComponentInterface, MutableAriaAttribute {
     /**
      * If `true`, the user cannot modify the value.
      */
-    @Prop() readonly = false;
+    @Prop({ reflect: true }) readonly = false;
 
     /**
      * If `true`, the user must fill in a value before submitting a form.